Breaking up carousel and fetching logic.
