+++ src/modules/browse/components/ShelfBrowseCarousel/styles.css
@@ -1,0 +1,234 @@
+.shelf-browse,
+.shelf-browse > header {
+  align-items: baseline;
+}
+
+.shelf-browse {
+  display: grid;
+  gap: 1rem;
+  grid-template-areas:
+    "header header header"
+    "carousel carousel carousel"
+    ". return .";
+  grid-template-columns: 1fr auto 1fr;
+  grid-template-rows: auto;
+  order: 3;
+  padding: 1.5rem;
+}
+
+@media (min-width: 640px) {
+  .shelf-browse {
+    grid-template-areas:
+      "header return"
+      "carousel carousel";
+    grid-template-columns: 1fr auto;
+  }
+}
+
+.shelf-browse > header {
+  grid-area: header;
+  font-size: 0.875rem;
+}
+
+.shelf-browse-carousel {
+  display: grid;
+  gap: 1rem;
+  grid-area: carousel;
+  grid-template-areas: "prev list next";
+  grid-template-columns: auto 1fr auto;
+  grid-template-rows: auto;
+}
+
+.shelf-browse-carousel > button svg {
+  color: var(--search-color-blue-500);
+}
+
+.shelf-browse-carousel > button:disabled svg {
+  color: var(--search-color-grey-500);
+}
+
+.shelf-browse-carousel > button:not(:disabled):hover svg {
+  color: var(--ds-color-blue-400);
+}
+
+.shelf-browse-carousel > button:first-of-type {
+  grid-area: prev;
+  text-align: right;
+}
+
+.shelf-browse-items {
+  display: grid;
+  column-gap: 1rem;
+  font-family: "noto-sans-semicondensed", sans-serif;
+  font-size: 0.75rem;
+  grid-area: list;
+  grid-template-rows: 1fr auto;
+  position: relative;
+}
+
+.shelf-browse-item {
+  grid-row: 3 / 2;
+}
+
+.shelf-browse-item.animation-out {
+  animation: 0.25s ease-in-out fadeOut forwards;
+}
+
+.shelf-browse-item.animation-out-next {
+  animation:
+    0.25s ease-in-out fadeOut forwards,
+    0.25s ease-in-out slideOutFromLeft forwards;
+}
+
+.shelf-browse-item.animation-out-previous {
+  animation:
+    0.25s ease-in-out fadeOut forwards,
+    0.25s ease-in-out slideOutFromRight forwards;
+}
+
+.shelf-browse-item.animation-in {
+  animation: 0.25s ease-in-out fadeIn forwards;
+}
+
+.shelf-browse-item.animation-in-next {
+  animation:
+    0.25s ease-in-out fadeIn forwards,
+    0.25s ease-in-out slideInFromRight forwards;
+}
+
+.shelf-browse-item.animation-in-previous {
+  animation:
+    0.25s ease-in-out fadeIn forwards,
+    0.25s ease-in-out slideInFromLeft forwards;
+}
+
+@media (prefers-reduced-motion) {
+  .shelf-browse-item[class*="animation-"] {
+    animation: none;
+  }
+}
+
+.shelf-browse-item > a,
+.shelf-browse-item > a > dl {
+  height: 100%;
+}
+
+.shelf-browse-item > a {
+  color: inherit;
+  display: block;
+}
+
+.shelf-browse-item-current {
+  grid-row: 3 / 1;
+}
+
+.shelf-browse-item-current > a {
+  border: 4px solid var(--search-color-blue-500);
+}
+
+.shelf-browse-item dl {
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.shelf-browse-item .this-item {
+  align-self: baseline;
+  background: var(--search-color-blue-500);
+  color: white;
+  font-size: 0.5rem;
+  letter-spacing: 1px;
+  margin-left: calc(-0.75rem - 1px);
+  padding: 0.5rem 1.25rem 0.5rem 0.5rem;
+  text-transform: uppercase;
+  clip-path: polygon(0% 0%, 100% 0, calc(100% - 0.75rem) 50%, 100% 100%, 0% 100%);
+}
+
+.shelf-browse-item dd {
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+}
+
+.shelf-browse-item .this-item,
+.shelf-browse-item .item-term-title,
+.shelf-browse-item .item-term-callnumber_browse {
+  font-weight: 700;
+}
+
+.shelf-browse-item .item-term-published_year,
+.shelf-browse-item .item-term-callnumber_browse {
+  -webkit-line-clamp: 1;
+}
+
+.shelf-browse-item .item-term-title {
+  color: var(--search-color-blue-400);
+  text-decoration: underline;
+}
+
+.shelf-browse-item > a:hover .item-term-title {
+  text-decoration-thickness: 2px;
+}
+
+.shelf-browse-carousel > button:last-of-type {
+  grid-area: next;
+  text-align: left;
+}
+
+.shelf-browse-carousel-return {
+  grid-area: return;
+}
+
+@keyframes slideInFromLeft {
+  from {
+    transform: translateX(-1rem);
+  }
+  to {
+    transform: translateX(0);
+  }
+}
+
+@keyframes slideInFromRight {
+  from {
+    transform: translateX(1rem);
+  }
+  to {
+    transform: translateX(0);
+  }
+}
+
+@keyframes slideOutFromLeft {
+  from {
+    transform: translateX(0);
+  }
+  to {
+    transform: translateX(-1rem);
+  }
+}
+
+@keyframes slideOutFromRight {
+  from {
+    transform: translateX(0);
+  }
+  to {
+    transform: translateX(1rem);
+  }
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes fadeOut {
+  from {
+    opacity: 1;
+  }
+  to {
+    opacity: 0;
+  }
+}
